cmd/snap-bootstrap/partition: add Disk, methods to cross-check mountpoints
These methods are meant to be used by snap-bootstrap, primarily to take an
already mounted partition, and use that to find the underlying disk, then to
find an unmounted partition from the same disk, with the specified label so that
we can mount that one.

We also have the use case that a given mountpoint might be a decrypted mapper
device, in which case we need some special logic to dig through sysfs and udev
to find the underlying encrypted device that the decrypted mapper device maps
to.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>