o/devicestate: change how current system is reported for different modes
Change how current system is determined in different modes. In run mode, we use
the seeded-systems information that is updated in mark seeded task.

In recover and install modes, we are effectively seeding an ephemeral system, so
no such information is recorded. Use the modeenv to identify that recovery
system that booted and use it as the current system. However, in install mode,
report nothing for now. Later we should return a conflict error.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>