tests: reload root's systemd --user after snapd tests
Recently we realized that some tests leave systemd/user configuration
files in /etc that are normally not cleaned up by snapd under any
circumstances. Despite this we still, occasionally, see some failures
related to session-agent tests.

With extra debugging we understand that the problem is caused by the
fact that loaded units stay in systemd --user's memory (for the root
user) between tests. Unlike the session of the test user, we are not
performing a full cleanup, as that would nuke our SSH connection from
spread.

This patch fixes the affected tests to reload systemd --user running for
root, in the appropriate restore sections.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>