tests: port interfaces-dbus to session-tool
This test did some weird hacks that I removed:
 - it faked a session bus on systems where it is unsupported
 - it spawned DBus and left it running
   - dbus.sh always leaks a process behind
 - it faked an X11 server to cause python to spawn spawn another bus!
 - it used an endless loop hoping for the service to respond
   - replaced with retry-tool

With this patch the test prepare/restore logic is more straightforward
and cannot leak anything behind as session-tool terminates all the
entire user slice.

In addition, this test could pass on core20 if we exclude the first
interaction, which uses unconfined user to check if the service is
"alive". On core systems the DBus interface does not allow unconfined
bus access so this check would always fail. I chose not to change that
for now.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>