o/devicestate: verify conflicts with seeding when system action is requested
When we get a request for a system action, verify that the device is seeded or
at least that the seeding change has finished.

We cannot execute the action sooner, because the list of allowed actions for the
current system depends on the mode and the system state. Eg. a fully seeded
system supports 'run'/'recover'/'install' actions for itself, or install for other
system seeds. However, when the seeding fails, one can only execute an 'install'
action for the system.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>