osutil/disks: support IsDecryptedDevice for mountpoints which are dm devices
This will allow us to cross-check that partitions mounted from the initramfs
that are really decrypted mapper devices that corespond to physical encrypted
partitions.

Also add some additional tests for this new functionality, including that we
correctly error out when the device is expected to be a decrypted device, but is
actually a physical partition.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>