overlord: show what manager throws a state ensure error
When we install UC20 we get a lot of state.ErrNoState errors. In
general the state engine error reporting can be unhelpful if the
errors generic. To improve the situation a bit this commit adds
information about what manager is throwing the error. Eventually
we should ensure we don't return "naked" ErrNoState or similar
generic errors all the way up without decorating them with more
information.