state: log task errors in the journal too
When we see failures in snapd operations like installing a UC20
system or installing a single snap we need to ask for two logs
currently. The journal of snapd and the snap changes output with
the details for the failed change.

This commit tries to improve this situation by logging task
errors into the snapd journal as well. This allows users to just
sent the `journalctl -u snapd` output for bugreports in most
cases.