packaging/arch: update PKGBUILD to match one in AUR
Update the PKGBUILD to match what is currently available in AUR.

The changes are mostly minor or cosmetic. A larger tweak for the go flags makes
sure that quoting is corrrect when passing nested flags eg.
`-ldflags "-extldflags='-flag -flag'"`.

Adopt a check for static binaries we already have in the RPM packaging.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>