tests: detect signs of crashed snap-confine
During a specific phase of the construction of the mount namespace,
snap-confine creates a directory matching /tmp/snap.rootfs_*. Normally
by the time snap-confine passes the baton to snap-exec, this directory
is gone. It can only stay behind if snap-confine crashes at a particular
point.

Detecting this directory may shed some light on a particular test
failure, where /tmp/snap.rootfs_* was a mount point (which is expected)
and was failed to be removed by the test prepare/restore logic.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>