cgroup: test session -> system bus fallback
When snapd wants to track an application it is launching, it is asking
systemd to create a transient scope unit. Normally the session systemd
is used (aka systemd --user) but if that bus is not available and if the
user is root, then the logic automatically falls back to the system bus.

This allows tracking applications that live outside of any session, as
long as they run as root. One concrete example is all of the snap hooks,
that are invoked by snapd.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>