o/configstate/configcore: handle on-demand console-conf/(serial-)getty instaces
Getty and console-conf are both template services. Typically in a UC20 image
only getty@tty1 is defined as a side effect of being 'wanted' by the
getty.target.

When a tty2 (or other terminal) becomes active, console-conf or getty will be
spawned there accordingly. However, on those terminals, the getty/console-conf
unit is started on demand, and technically does not exist when stopped. Thus,
globs like `getty@*` will not pick up and restart those services. We explicitly
have to act on the active console-conf instances to trigger the right behavior.

Similar handling has to be applied to serial-getty@ and serial-console-conf@
services.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>