testutil/exec.go: set PATH after running shellcheck
If we set this before calling shellcheck, if shellcheck is a snap, then we could
have issues mocking specifically systemctl, where shellcheck as a snap tries to
use `systemctl is-system-running` if there is a system key mismatch, and this
will call our mocked systemctl which leads to undesirable and very confusing
unit test failures.

Longer term we may want to consider not setting PATH in this manner as there
have been other similar bugs where some system tool is a snap and it affects
unit tests in a confusing way.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>