snap/container.go: don't import snapdir, squashfs; use register format w/ init()
Here we disentangle snap/container.go from importing snapdir and squashfs so
that in a followup PR we can add an option struct to Install() that then is used
in snapdir / squashfs implementations of Install().

We do this by having container.go export a register format function that is
called during init() by the snapdir and squashfs packages. This required some
refactoring to make snapdir work with a magic identifying function the way that
squashfs currently does.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>