snap/many: support InstallOptions + MustCopy for Install()
This will allow us from snapstate to decide how a snap file should be installed,
if it should be symlinked or if it should be always copied, etc. Eventually we
should also introduce logic to snapstate to decide if it should be linked to
remove the overlayfs specific logic from here where it is kind of awkward to
see.

Also add tests for the squashfs and snapstate implementations.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>