tests: port interfaces-many-core-provided to tests.session
This test, surprisingly, spawns and leaves behind the XDG document
portal. By exercising all interfaces, one by one, it invokes a app that
has the desktop plug declared. This causes "snap run" to activate the
document portal over D-Bus.

On systems without D-Bus session bus available, it will also spawn and
leak a dbus-daemon. On systems with the D-Bus session bus available, it
will just leak the document portal service.

As a solution, if possible, run all the tests via tests.session, inside
the session of the test user, that is automatically torn down along all
the processes in it. On systems where we cannot use tests.session, we
run all but the "desktop" test app directly as root. The desktop test
app is left out and not tested on systems that, arguably, are not very
desktop like. This leaves out Amazon Linux 2, CentOS 7 and also Ubuntu
14.04 which is not supported for desktop applications.

With this patch, there is only one remaining test leaking dbus-daemon,
the one testing xdg-settings.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>