tests: port snap-routine-portal-info to session-tool
This test is rather straightforward. The only interesting highlight is
the use of user session service to run the "sleep 1h" command and a
rather non-trivial way to extract the PID of the sleep process, due to
the stack of a snap command and session-tool (we just ask systemd). This
method of running background tasks is more resilient as they are
properly shut down by session-tool --restore -u test *and* they don't
hang spread *and* this also kills any helper processes that may have
been started.

This change removes one more leaking dbus-daemon, namely that from
test-snapd-desktop.cmd which causes desktop portal activation.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>