tests: reload systemd --user for root, if present
If the root user has a systemd --user instance then ask it to reload.
This prevents tests from leaking user-session services that stay in
memory but are not present on disk, or have been modified on disk, as is
common with tests that use snaps with user services _or_ with tests that
cause installation of the snapd.session-agent.service unit via re-exec
machinery.

This is done AHEAD of the invariant checks as it is very widespread
and fixing it in each test is not a priority right now.

Note that similar treatment is not required for the "test" user as
correct usage of session-tool ensures that the session and all the
processes of the "test" user are terminated.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>