snap/naming: add ParseSecurityTag and friends
We have a need to parse security tags and extract specific fields,
like snap name, in a reliable way. This patch introduces a new function,
ParseSecurityTag, along with a group of related interfaces
ParsedSecurityTag, ParsedAppSecurityTag and ParsedHookSecurityTag.

Together they piggy back on the work that was already done for
ValidateSecurityTag, to return the right data at the right time.
At the same time the aforementioned function reduces to a trivial
call to ParseSecurityTag.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>