tests: clean state of snapd.session-agent.socket after snapd purge
When a socket file that was active is removed from disk, reloading
systemd does not stop the socket. The socket is stuck in a limbo state
described by systemd in the following way:

    # systemctl --user status snapd.session-agent.socket
    ● snapd.session-agent.socket
       Loaded: not-found (Reason: Unit snapd.session-agent.socket not found.)
       Active: active (listening) since Fri 2020-05-29 09:39:55 UTC; 8min ago
       CGroup: /user.slice/user-0.slice/user@0.service/snapd.session-agent.socket

    May 29 09:39:55 may290939-310559 systemd[1269]: Listening on REST API socket for snapd user session agent.
    May 29 09:47:45 may290939-310559 systemd[1269]: snapd.session-agent.socket: Socket unit configuration has changed while unit has been running, no o

This corresponds to the TODO both the snap-mgmt script and the Debian
postrm script, which talks about management of state of various systemd
--user instances on the system. Currently this is not done.

To avoid the test failures that this is causing, in while performing
test cleanup, stop the socket managed by systemd --user (for the root
user), and start it again, ignoring errors if it does not exist.

Note that similar handling is not required for systemd --user of the
test user as the entire session is torn down, along with all state,
involving systemd --user state and all the sockets it describes.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>