tests: use tests.session exec
According to the tools agreement, executing a command in a session needs
to be a proper sub-command.

This is an almost entirely automatic change. The result was hand-edited
to replace "tests.session ... exec -p ARG" with "tests.session ... -p
ARG exec":

    find -type f -exec perl -pi -e 's/tests.session -u (\S+) (?!prepare|restore|exec)/tests.session -u \1 exec \2/g' {} \;

The use of perl is to allow for negative lookahead, which sed does not
support.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>