tests: port document-portal-activation to session-tool (#8733)
* tests: port document-portal-activation to session-tool

There are a number of usual changes, move to session-tool, not leaking
dbus-daemon, running a session service as the test user. That last
change mainly helps to make test cleanup reliable. I also did three
extra changes along the way:

 - The check_stderr helper no longer checks cgroup type each time.
 - The helper variables were removed, this makes some commands longer
   but more understandable and less quirky to quoting and shell
   expansion issues.
 - The variant where session bus is unavailable is now tested last,
   with session bus socket stopped and with environment override
   that stops snap run from trying to talk to a non-existent socket.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* tests: restore more strict error check

The test previosuly verified that the *only* thing present is the
warning about cgroupv2 support. This patch restores that strictness.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>