cmd/snap-confine: add support for libc6-lse
We've been alerted that snapd doesn't operate on new aarch64 virtual
machines, coupled with the new libc6-lse package. LSE stands for Large
Systems Extensions. Here it is materialized as a libc with support
libraries in a sub-directory "atomics", underneath the arch triplet.

	$ dpkg -L libc6-lse
	[... irrerlevant parts hidden]
	/lib/aarch64-linux-gnu/atomics/ld-linux-aarch64.so.1
	/lib/aarch64-linux-gnu/atomics/libBrokenLocale.so.1
	/lib/aarch64-linux-gnu/atomics/libanl.so.1
	/lib/aarch64-linux-gnu/atomics/libc.so.6
	/lib/aarch64-linux-gnu/atomics/libdl.so.2
	/lib/aarch64-linux-gnu/atomics/libm.so.6
	/lib/aarch64-linux-gnu/atomics/libnsl.so.1
	/lib/aarch64-linux-gnu/atomics/libnss_compat.so.2
	/lib/aarch64-linux-gnu/atomics/libnss_dns.so.2
	/lib/aarch64-linux-gnu/atomics/libnss_files.so.2
	/lib/aarch64-linux-gnu/atomics/libnss_hesiod.so.2
	/lib/aarch64-linux-gnu/atomics/libnss_nis.so.2
	/lib/aarch64-linux-gnu/atomics/libnss_nisplus.so.2
	/lib/aarch64-linux-gnu/atomics/libpthread.so.0
	/lib/aarch64-linux-gnu/atomics/libresolv.so.2
	/lib/aarch64-linux-gnu/atomics/librt.so.1
	/lib/aarch64-linux-gnu/atomics/libthread_db.so.1
	/lib/aarch64-linux-gnu/atomics/libutil.so.1

This patch changes the snap-confine and snap-update-ns apparmor profiles
to allow optional linking with those binaries.

Note that I specifically do not change the base apparmor template, which
precludes from using libc6-lse in a base snap. I would like to follow up
with security before we do that.

Fixes: https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1881350
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>