boot/bootstate16.go: clean snap_try_* vars when not in Trying status too
This is needed because the uc18/uc16 bootloader implementations do not actually
do this, and instead snapd needs to reset the value.

Also modify the bootenv bootloader logic simulation to not clean this anymore,
just as the real bootloaders do. 

This all requires some slight changes to the test logic, which since they are 
checking the BootVars map directly will see some additional empty keys in the 
map as markSuccessful will now write those empty keys. Eventually we should port
these tests to use GetBootVars and not use the map directly since empty 
string keys show up the same as missing keys with GetBootVars and that would 
have prevented needing to change these tests at all.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>