data/selinux, packaging/fedora: exclude fwupd policy on RHEL7
The fwupd_search_cache() interface template is not defined on RHEL7.

In theory we should be able to use the `optional_policy(..)` block to capture
that, but based on investigation, the optional block is not correctly handled
when expanding the policy in kernel policy language. After m4 preprocessing is
done, the interface is not expanded and causes a compilation error further down
the stack. Based on feedback from #selinux, this works in CIL (should we ever
use it).

Make the relevant policy bit ifdef()'ed based on the distro.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>