sanity: check for unsynchronized real time clock
Some popular core devices do not have a real-time clock and boot with an
arbitrary date baked into the boot loader or initrd. Snapd relies on
correct clock for validating assertions and checking SSL certificates.

To avoid confusion, fail early and loud, when the clock is before a
canned date.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>