New skip binary to stip tests in an easy way
This new SKIP binary allows to simplefy the way we check and exit from
tests.

The idea is that SKIP receives a line, which determines if the phase has
to be skept. In case it evaluation is true, then the phase in exited
with 0. In case the SKIP does not receive any parameter, it will check
if the evaluation has been already done and in this case it will exit
the current phase.