sandbox/cgroup: remove redundant pathOfProcPidCgroup
The cgroup package already contains the identical ProcPidPath which
also has the benefit of being testable without a special mock function.

There's a tiny change to tests, to phase out the custom mock function
MockPathOfProcPidCgroup. In a follow-up we should probably make
ProcPidPath private as it seems otherwise unused outside of this
package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>