snap: add snap export meta-data
Snaps can export content to the host. Exports are described by a map of
paths, as seen from a fixed location on the host, to files and
directories relative to the mount point of a specific snap revision.

Currently exports cannot be declared by YAML and are reserved for the
snapd snap.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>