sysconfig/cloudinit: add RestrictCloudInit
RestrictCloudInit will implement the cloud-init specific bits of mitigation
against CVE-2020-11933, insofar as it will disable importing of cloud-init
NoCloud datasources from arbitrary filesystems that an attacker could control
and put malicious cloud-init data on as well as always limit the datasource for
cloud-init to use on subsequent boots to the detected datasource.

We have to parse the cloud-init status.json file in order to determine what
datasource was used because while the cloud-init status does give us the same
information, it is in a less useful format that is harder to parse than the JSON
here, and the JSON here is in a stable v1 structure that can be relied upon.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>