build-aux/snapcraft.yaml: use focal fontconfig too
This is a complete and total hack, but it does actually seem to work for the
snapd snap at least. Still need to port this hack to the core snap and snapd deb
somehow but otherwise it works.

Fixing the spread tests temporarily by forcing them to use the snapd snap from
the branch.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>