sandbox/cgroup: allow discovering PIDs of given snap
The main function PidOfSnap, returns information about the processes of
a given snap, including the application or hook each process belongs to.

The returned information is inherently volatile, as processes can die
and fork at will. To use correctly it needs to be paired with
appropriate locking. The main consumer will be the snap manager, which
will use it around refreshes.

The final usage depends on the both the snap lock and the snap run
inhibition lock, which is critical for consistency of the result.
Holding the snap lock, perform the measurement, depending on the result
either release the snap lock and do nothing new or obtain the snap run
inhibition lock and release the snap lock. This method guarantees that
we can reliably make a decision not on the particular pid values but on
the set of applications and hooks that are running - exactly what
refresh app awareness was designed to need.

The actual logic is relatively simple and involves scanning cgroups for
directories named snap.$SNAP_INSTANCE_NAME.*, each directory has a
cgroup.procs file that contains a snapshot of the PIDs inhabiting that
cgroup at a given time. All cgroup directories that belong to the snap
in question are scanned and aggregated.

Both cgroup v1 and cgroup v2 modes are supported. In v1 mode we only
scan the systemd tracking cgroup. In v2 mode there is only one directory
with unified cgroup so everything is canned.

There's a collection of unit tests. Spread tests are bundled with the
complete feature, where we measure application tracking in detial.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>