tests/main/xdg-open-compat: backup and restore original xdg-open
The test overwrites and then removes its copy of /usr/bin/xdg-open, but never
restores the original file. This would break the tests that follow that rely on
the behavior of xdg-open. Specifically snap-handle link was broken like so:

2020-06-15 19:35:11 Error executing google:ubuntu-16.04-64:tests/main/snap-handle-link (jun151913-234107) :
-----
...
+ tests.session -u test exec env DISPLAY=:placeholder xdg-open snap://package
env: 'xdg-open': No such file or directory
grep error: pattern not found, got:
-----

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>