boot/bootstate20: refactor genericInitramfsSelectSnap
This makes the function simpler by returning one less return value which was
redundant, in that the expectFallback bool really is only true when second!=nil,
so the caller can just make that comparison directly.

Also refactor the logic to be more straight-forward, in that we just check for
each error condition in order before returning the final case of a proper try
snap. This is easier to follow and reason about than the previous nested if's
and switch statements. Thanks to @bboozzoo for the suggestion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>