secboot: implement key resealing
The key to the encrypted data partition is sealed to a number of
different system measurements including the kernel command line, the
snap system "epoch", the snap model, and the EFI binaries in the
boot chain. When one of these parameters is changed in an update,
unlocking the encrypted data partition will no longer be possible
unless we update the protection profile for the existing sealed key.
This change adds the call to do the resealing operation based on
the new system parameters.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>