snapstate: fix autorefresh from classic->strict
We got a bugreport that going from classic->strict in a refresh
does not work. This commit fixes this issue.

Back in c6117368b75 code was added in validateFlagsForInfo() to
check that it is not possible to do:
```
$ snap install --classic test-snapd-tools
```
i.e. this resulted in an error with this check.

However in 9e3b3a43f32 this was changed so that it is now possible
to do such an install. If this happens the --classic option is
simply ignored.

This commit now changes the code from 9e3b3a43f32 to also apply
to refreshes, not just installs.  This fixes refreshes that go
from classic->strict. This unblocks multipass (LP: #1883538).