sysconfig: add CloudInitRestrictOptions w/ ForceDisable, err on errored/enabled
This means that a caller will have to know whether or not they want to forcibly
disable cloud-init when in the errored or enabled states, as since these are not
"steady" states, we don't want to automatically disable cloud-init here.

Also re-order the fields in CloudInitRestrictionResult to match the order they
are introduced in the doc-comment.

Finally, use dirs.GlobalRootDir as the target dir for DisableCloudInit when
called from RestrictCloudInit. This is to avoid passing in a TargetRootDir via
some sysconfig.Options to the function.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>