tests/lib/tools: apply linger workaround when needed
We would always a workaround for /var/lib/systemd/linger, however some
distros (eg. CentOS 8, Fedora 31/32) seem to work and manually adding
StateDirectory to service seems to break things. In particular on Centos8, it
makes systemd trip over the SELinux the policy which expects a dynamic
transition to happen.

Try to selectively apply a workaround on systems where enabling-linger does not
work on a first try.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>