tests/lib/nested.sh: use vfat usb drive in qemu for assertion disk
The current implementation of automatically importing the system-user assertion
did not work at all, as snap auto-import failed with "no medium found" when
attempting to mount the drive. It's unclear how to make that work with debugfs
and an ext4 formatted drive that does not use partitions, but this method does
work instead.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>