spread.yaml: add qemu-nested backend with 8G mem, add to nested suites
The memory for the qemu backend needs to be at least 5G because the nested VM's
are created with 4G of memory, so just re-using the current 2G of memory
specified for the qemu backend will lead to the VM's not being able to be
booted. It's also unreasonable to expect that folks just wanting to run 2 qemu
spread tests locally need 16G+ of memory, so we define a new backend just
specifying 8G of memory.

Also add 16.04 and 18.04 to the nested manual suites.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>