tests/lib/nested.sh, spread.yaml: use backend specific opts for RAM, machine
For qemu-nested, we can't use acceleration because it doesn't support nesting,
so we omit that option when using qemu-nested.

Additionally, since qemu-nested is likely to be run on developer machines that
may not have a significant amount of RAM, we reduce the amount of RAM for the
guest VM to 2GB, and the RAM for the host VM to 4G.

Finally, the -usb option to qemu was included in the PARAM_MONITOR, but it
doesn't really have anything to do with monitoring, so move it to it's own
parameter variable.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>