tests/lib/prepare.sh: for classic, don't fail if core snap is still there
Instead we just refresh the core snap if it is there as we cannot remove the
core snap when we have snapd installed as it is always required.

If the core snap is not there, then just install it as before.

This is necessary because when re-using nested systems specifically, the core
snap is not properly cleaned up from something in the restore section most
likely.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>