interfaces: add cups companion interface to cups-control
On systems where the slot is provided by an app snap, the cups interface is
the companion interface to the cups-control interface. The design of these
interfaces is based on the idea that the slot implementation (eg cupsd) is
expected to query snapd on if the cups-control interface is connected or not
and the print service will mediate admin functionality (ie, the rules in
these interfaces allow connecting to the print service, but do not implement
enforcement rules; it is up to the print service to provide enforcement).

For now we limit this to app-providing slot implementations since there
are currently no classic systems with a mediating cupsd. We can revisit
this when it becomes available.