sandbox/cgroup: add TODO about hooks and session bus
Discussion with Maciej resulted in the realisation that when snap hook
is executing it should not be tracked through the session bus of the
root user, even though this is possible and may work, but instead
through the system bus.

Add a TODO note to do this in a follow-up.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>