sandbox/cgroup: use more strict cgroup membership test
The code used to check if the systemd scope name we selected is among
the effective path in the tracking cgroup hierarchy. Maciej indicated
that a more strict check should be sufficient, namely that the unit name
is the suffix of the path. After some consideration I don't see any
reason for the laxness of the original check.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>