sandbox/cgroup: extend SnapNameFromPid with tracking cgroup data
The function SnapNameFromPid is used by `snap routine ...` to determine
if a given process is a snap. The function used to look for the
freezer cgroup but this is no longer used in cgroupv2 world.
Instead, it will now look for the new tracking cgroup, if one is available
and if not, fall back to the freezer cgroup.

In practice, this will soon allow `snap routine portal-info` to work
on Fedora 31+. This is not yet possible as we do not track applications
just yet.

This patch was extracted out of the larger feature branch.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>