tests: stop root's session dbus.service
The cgroup-tracking test runs various scenarios to examine how cgroup
tracking behaves in practice. Once case involves the root user. On some
systems the root user has a session with a session bus that can be
accessed by the tracking feature. On those systems that bus is
socket-activated and stays behind after the test executes.

Due to the limitations in tests.session ability to restore the session
for the root user, handle dbus.service via a special case in those
specific tests.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>