tests/lib/nested.sh: split out start_nested_core_vm into helper funcs
We now are able to create/start a nested VM with start_nested_core_vm, then stop
it with force_stop_nested_vm which stops the systemd unit for running QEMU, and
then re-start the VM with different arguments using start_nested_core_vm_unit
without running into issues like prepare_ssh failing to run twice because the
test user is already there, or deleting/recreating the VM disk image file, etc.

We also block waiting for the systemd unit to become active, because if we have
a catastrophic failure in running qemu in the systemd unit, we don't want to
wait for SSH to become available because it never will, so this will make the
logs more obvious that the nested-vm systemd unit is down.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>