tests/core/gadget-update-pc: port the test to UC20
Enable the test on UC20. We need to adjust how the gadget.yaml gets generated
because of the changes in naming/types/roles between UC16/UC18 and UC20. Extend
the test to verify that managed boot config is preserved and that gadget assets
are correctly deployed to the seed partition.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>