tests: rename user-tool to user-state, fix --help
In accordance with the tooling naming agreement, user-tool becomes
user-state and is invoked with the full path. It also no longer crashes
when invoked without arguments and responds to -h and --help.

There's also a trivial test to ensure it works enough to print the
help message on all the systems.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>